requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Werkzeug==2.2.3
imaplib2==3.6
email-validator==2.1.0.post1
python-dotenv==1.0.1
PyMuPDF==1.23.20
flask-migrate
python-dateutil
gunicorn
